+++ app/Repositories/ContractorsRepository.php
@@ -6,7 +6,7 @@ use Hash;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Illuminate\Http\Request;
 use App\Repositories\ContractorsRepositoryInterface;
-use App\Models\{ Contractor };
+use App\Models\Contractor;
 
 class ContractorsRepository implements ContractorsRepositoryInterface
 {
@@ -17,8 +17,10 @@ class ContractorsRepository implements ContractorsRepositoryInterface
         $this->model = $contractor;
     }
 
-    public function all($search = '')
+    public function all($search = '', $config = [])
     {
+        $shouldPaginate = isset($config['paginate']) ? $config['paginate'] : true;
+
         if ($search) {
             $query = Contractor::
                 where('company', 'like', "%".$search."%")
@@ -31,21 +33,29 @@ class ContractorsRepository implements ContractorsRepositoryInterface
                     $query->where('cities.name', 'like', $search);
                 });
         } else {
-            $query = Contractor::query();;
+            $query = Contractor::query();
         }
 
-        return $query
-            ->orderBy('company', 'asc')
-            ->paginate(30);
+        $query->orderBy('company', 'asc');
+
+        if($shouldPaginate) {
+            $result = $query->paginate(30);
+        }else{
+            $result = $query->get();
+        }
+        
+        return $result;
     }
 
     public function create(Request $request)
     {
+        ContractorsValidations::validateOnCreate($request);
         return $this->save($request);
     }
 
     public function update(Request $request, $id)
     {
+        ContractorsValidations::validateOnEdit($request, $id);
         return $this->save($request, $id);
     }
 
@@ -53,15 +63,12 @@ class ContractorsRepository implements ContractorsRepositoryInterface
     {
         $is_new = ! $id;
 
-        // en esta zona obtenemos los datos o preparamos las variables para asignarle lo que se envió en los formularios
         if($is_new){
             $contractor = $this->blueprint();
-            $request->validate(Contractor::$saveValidation);
         }else{
             $contractor = $this->find($id);
-            $request->validate(Contractor::$updateValidation);
         }
-        // despues el guardado de los datos
+
         $contractor->fill($request->all());
         $contractor->save();
 
@@ -73,10 +80,6 @@ class ContractorsRepository implements ContractorsRepositoryInterface
         $is_obj = is_object($id_or_obj);
         $contractor = ($is_obj) ? $id_or_obj : $this->model->find($id_or_obj);
 
-        if ($contractor) {
-            $contractor->first();
-        }
-
         if (!$contractor) {
             throw new ModelNotFoundException("Contractor not found");
         }
@@ -86,23 +89,20 @@ class ContractorsRepository implements ContractorsRepositoryInterface
 
     public function delete($id)
     {
-        $contractor =
-            $this->model
-                ->where('id', '>', '0') // to avoid deleting seed items
-                ->find($id);
-
-        if ($contractor) {
+        $contractor = $this->model->find($id);
+        
+        if ($contractor && $this->canDelete($id)) {
             $contractor->delete();
         }
 
-        // return recently deleted object to be used if needed after operation
-        // the object may or may not exists
         return $contractor;
     }
 
-    /**
-     * Return the blueprint of the model including translation elements
-     */
+    public function canDelete($id) 
+    {
+        return true;
+    }
+
     public function blueprint()
     {
         $contractor = new Contractor;
